requirements.txt
streamlit
pandas
gspread
oauth2client
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
python-dotenv
PyPDF2
python-docx
openpyxl
python-pptx
Pillow
reportlab
flask
flask-cors